requirements.txt
adagio==0.2.4
aiofiles==0.8.0
aiohttp==4.0.0a1
aiosignal==1.3.1
aiosqlite==0.17.0
alembic==1.8.1
altair==4.2.0
ansi2html==1.8.0
antlr4-python3-runtime==4.9.3
anyconfig==0.10.1
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.2
asttokens==2.1.0
async-generator==1.10
async-lru==1.0.3
async-timeout==3.0.1
attrs==21.4.0
autopage==0.5.1
autoviz==0.1.58
Babel==2.11.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
binaryornot==0.4.4
bleach==5.0.1
blinker==1.5
bokeh==2.4.3
boto3==1.26.16
botocore==1.29.16
Brotli==1.0.9
build==0.9.0
CacheControl==0.12.11
cachetools==4.2.4
cachy==0.3.0
catboost==1.1.1
category-encoders==2.5.1
# certifi @ file:///croot/certifi_1665076670883/work/certifi
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
chart-studio==1.1.0
cleo==1.0.0a5
click==8.0.4
cliff==4.1.0
cloudpickle==2.2.0
cmaes==0.9.0
cmd2==2.4.2
codecov==2.1.12
colorcet==3.0.1
colorlog==6.7.0
colour==0.1.5
commonmark==0.9.1
contourpy==1.0.6
convertdate==2.4.0
cookiecutter==2.1.1
coverage==6.6.0b1
crashtest==0.3.1
cryptography==38.0.3
cycler==0.11.0
Cython==3.0.0a11
daal==2021.7.0
daal4py==2021.7.0
dash==2.7.0
dash-auth==1.4.1
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-cytoscape==0.3.0
dash-html-components==2.0.0
dash-table==5.0.0
#dask==2022.11.1
databricks-cli==0.17.3
dataclasses==0.6
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.6
distlib==0.3.6
#distributed==2022.11.1
docker==6.0.1
dtreeviz==1.4.0
dulwich==0.20.50
dynaconf==3.1.11
emoji==2.2.0
entrypoints==0.4
evidently==0.2.0
exceptiongroup==1.0.4
executing==1.2.0
explainerdashboard==0.4.0
fairlearn==0.8.0
fastapi==0.66.1
fastjsonschema==2.16.2
featuretools==1.18.0
ffmpy==0.3.0
filelock==3.8.0
Flask==2.2.2
Flask-Compress==1.13
Flask-Cors==3.0.10
Flask-SeaSurf==1.1.1
flask-simplelogin==0.1.1
Flask-WTF==0.15.1
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fs==2.4.16
fsspec==2022.7.1
fugue==0.6.6
future==0.18.2
gevent==22.10.2
gitdb==4.0.10
GitPython==3.1.29
gradio==3.11.0
graphql-core==3.2.3
graphviz==0.20.1
greenlet==2.0.1
grpcio==1.51.0rc1
gunicorn==20.1.0
h11==0.12.0
HeapDict==1.0.1
hijri-converter==2.2.4
holidays==0.17.2
holoviews==1.14.9
html5lib==1.1
htmlmin==0.1.12
httpcore==0.15.0
httptools==0.5.0
httpx==0.23.1
hvplot==0.8.2
hydra-core==1.2.0
hyperopt==0.2.7
idna==3.4
ImageHash==4.3.1
imageio==2.22.4
imbalanced-learn==0.9.1
importlib-metadata==4.13.0
importlib-resources==5.10.0
iniconfig==1.1.1
interpret==0.3.0
interpret-core==0.3.0
ipykernel==6.17.1
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
isoduration==20.11.0
iterative-stratification==0.1.6
itsdangerous==2.1.2
jaraco.classes==3.2.3
jedi==0.18.2
jeepney==0.8.0
Jinja2==3.1.2
jinja2-time==0.2.0
jmespath==0.10.0
joblib==1.2.0
json5==0.9.10
jsonpointer==2.3
jsonschema==4.17.1
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-dash==0.4.2
jupyter-events==0.5.0
jupyter-lsp==1.5.1
jupyter-ydoc==0.3.0a1
jupyter_client==8.0.0a4
jupyter_core==5.0.0
jupyter_server==2.0.0rc7
jupyter_server_fileid==0.6.0
jupyter_server_terminals==0.4.1
jupyter_server_ydoc==0.4.0
jupyterlab==4.0.0a31
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterlab_server==2.16.3
kaleido==0.2.1
#kedro==0.18.3
#kedro-viz==5.1.1
keyring==23.11.0
kiwisolver==1.4.4
kmodes==0.12.2
korean-lunar-calendar==0.3.1
lightgbm==3.3.3
lime==0.2.0.1
linkify-it-py==1.0.3
llvmlite==0.38.1
locket==1.0.0
lockfile==0.12.2
lxml==4.9.1
m2cgen==0.10.0
Mako==1.2.4
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.1
mdurl==0.1.2
mistune==2.0.4
mlflow==2.0.1
mlxtend==0.21.0
more-itertools==9.0.0
moto==4.0.11.dev15
msgpack==1.0.4
multidict==4.7.6
multimethod==1.9
multiprocess==0.70.14
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.0rc1
nltk==3.7
notebook==7.0.0a9
notebook_shim==0.2.2
numba==0.55.2
numpy==1.22.4
oauthlib==3.2.2
omegaconf==2.2.3
optuna==3.0.3
orjson==3.8.2
outcome==1.2.0
oyaml==1.0
packaging==21.3
pandas==1.4.4
pandas-profiling==3.5.0
pandocfilters==1.5.0
panel==0.12.7
param==1.12.3a1
paramiko==2.12.0
parso==0.8.3
partd==1.3.0
patsy==0.5.3
pbr==5.11.0
pep517==0.13.0
percy==2.0.2
pexpect==4.8.0
phik==0.12.2
pickleshare==0.7.5
Pillow==9.3.0
pip-tools==6.10.0
pkginfo==1.8.3
platformdirs==2.5.4
plotly==5.11.0
plotly-resampler==0.8.2rc3
pluggy==1.0.0
pmdarima==1.8.5
poetry==1.2.2
poetry-core==1.3.2
poetry-plugin-export==1.2.0
prettytable==3.5.0
prometheus-client==0.15.0
prompt-toolkit==3.0.33
protobuf==3.20.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyamg==4.2.3
pyaml==21.10.1
pyarrow==10.0.1
pycaret==3.0.0rc4
pycparser==2.21
pycryptodome==3.15.0
pyct==0.4.8
pydantic==1.10.2
pydeck==0.8.0
pydub==0.25.1
Pygments==2.13.0
PyJWT==2.6.0
pylev==1.4.0
PyMeeus==0.5.11
Pympler==1.0.1
PyNaCl==1.5.0
pynndescent==0.5.8
pyod==1.0.6
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyperclip==1.8.2
pyrsistent==0.19.2
PySocks==1.7.1
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-json-logger==2.0.4
python-multipart==0.0.5
python-slugify==7.0.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pyviz-comms==2.2.1
PyWavelets==1.4.1
PyYAML==5.4.1
pyzmq==25.0.0b1
qpd==0.2.7
qtconsole==5.4.0
QtPy==2.3.0
querystring-parser==1.2.4
ray==2.1.0
regex==2022.10.31
requests==2.28.1
requests-toolbelt==0.9.1
responses==0.22.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==12.6.0
rope==0.21.1
s3transfer==0.6.0
SALib==1.4.6.1
schemdraw==0.15
scikit-image==0.19.3
scikit-learn==1.1.3
scikit-learn-intelex==2021.7.0
scikit-optimize==0.9.0
scikit-plot==0.3.7
scipy==1.8.1
seaborn==0.12.1
SecretStorage==3.3.3
selenium==4.2.0
semver==2.13.0
Send2Trash==1.8.1b0
shap==0.41.0
shellingham==1.5.0
six==1.16.0
skope-rules==1.0.1
skorch==0.12.1
sktime==0.13.4
slicer==0.0.7
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.44
sqlparse==0.4.3
stack-data==0.6.1
starlette==0.14.2
statsforecast==1.3.1
statsmodels==0.13.5
stevedore==4.1.1
strawberry-graphql==0.127.0
streamlit==1.15.0
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
tbats==1.1.1
tbb==2021.7.1
tblib==1.7.0
tenacity==8.1.0
tensorboardX==2.5.1
terminado==0.17.0
text-unidecode==1.3
textblob==0.17.1
threadpoolctl==3.1.0
tifffile==2022.10.10
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
toolz==0.12.0
toposort==1.7
tornado==6.2
tqdm==4.64.1
trace-updater==0.0.9
traitlets==5.5.0
treeinterpreter==0.2.3
triad==0.7.0
trio==0.22.0
trio-websocket==0.9.2
tune-sklearn==0.4.5
typeguard==2.13.3
types-toml==0.10.8.1
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
ua-parser==0.16.1
uc-micro-py==1.0.1
umap-learn==0.5.3
uri-template==1.2.0
urllib3==1.26.13
urllib3-secure-extra==0.1.0
uvicorn==0.17.6
uvloop==0.17.0
validators==0.20.0
virtualenv==20.16.7
visions==0.7.5
waitress==2.1.2
watchdog==2.1.9
watchgod==0.8.2
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
websockets==10.4
Werkzeug==2.2.2
widgetsnbextension==4.0.3
woodwork==0.19.0
wordcloud==1.8.2.2
wrapt==1.14.1
wsproto==1.2.0
WTForms==3.0.1
xgboost==1.7.1
xlrd==2.0.1
xmltodict==0.13.0
y-py==0.5.4
yarl==1.8.1
yellowbrick==1.5
ypy-websocket==0.5.0
zict==2.2.0
zipp==3.10.0
zope.event==4.5.0
zope.interface==5.5.2
